1 to 1 Saga Pattern

+++ PCRepairService/PCRepairService/Models/OutboxMessage.cs
@@ -2,7 +2,8 @@
 {
     public class OutboxMessage
     {
-        public int Id { get; set; }
+        //not used
+        public long Id { get; set; }
         public string? exchange {  get; set; }
         public string? messageType { get; set; }
         public string? content { get; set; }
